requirements.txt
altair==5.4.0
attrs==24.2.0
bcrypt==4.2.0
black==24.10.0
blinker==1.8.2
cachetools==5.5.0
captcha==0.6.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorcet==3.1.0
contourpy==1.3.1
cycler==0.12.1
extra-streamlit-components==0.1.71
filelock==3.16.1
flake8==7.1.1
fonttools==4.55.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.8
imageio==2.36.0
Jinja2==3.1.4
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
mypy-extensions==1.0.0
narwhals==1.5.5
networkx==3.4.2
numpy==2.1.0
opencv-contrib-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pdf2image==1.17.0
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.0
protobuf==5.27.3
pyarrow==17.0.0
pycocotools==2.0.8
pycodestyle==2.12.1
pydeck==0.9.1
pydicom==2.4.4
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
scikit-image==0.24.0
scipy==1.14.1
six==1.16.0
smmap==5.0.1
st-clickable-images==0.0.3
streamlit==1.37.1
streamlit-authenticator==0.3.3
sympy==1.13.1
tenacity==8.5.0
tifffile==2024.9.20
toml==0.10.2
torch==2.5.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
